Added exception filters and statuses

+++ api/src/features/users/users.controller.ts
@@ -8,11 +8,15 @@ import {
   Delete,
   HttpCode,
   HttpStatus,
+  BadRequestException,
   NotFoundException,
+  UseFilters,
 } from '@nestjs/common'
+import { isValidObjectId } from 'mongoose'
 import { UsersService } from './users.service'
 import { CreateUserDto } from './dto/create-user.dto'
 import { UpdateUserDto } from './dto/update-user.dto'
+import { MongoExceptionFilter } from '@app/libs/filters'
 
 @Controller('users')
 export class UsersController {
@@ -20,8 +24,9 @@ export class UsersController {
 
   // Creates a new user
   @Post()
-  async create(@Body() newUserData: CreateUserDto) {
-    return await this.usersService.create(newUserData)
+  @UseFilters(MongoExceptionFilter)
+  async create(@Body() createUserDto: CreateUserDto) {
+    return await this.usersService.create(createUserDto)
   }
 
   // Returns all users
@@ -33,6 +38,8 @@ export class UsersController {
   // Returns one user
   @Get(':id')
   async findOne(@Param('id') id: string) {
+    if (!isValidObjectId(id))
+      throw new BadRequestException('Incorrect id format')
     const user = await this.usersService.findOne(id)
     if (!user) throw new NotFoundException()
     return user
@@ -40,8 +47,10 @@ export class UsersController {
 
   // Updates a user and returns it
   @Patch(':id')
-  async update(@Param('id') id: string, @Body() updateUserData: UpdateUserDto) {
-    const updatedUser = await this.usersService.update(id, updateUserData)
+  async update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    if (!isValidObjectId(id))
+      throw new BadRequestException('Incorrect id format')
+    const updatedUser = await this.usersService.update(id, updateUserDto)
     if (!updatedUser) throw new NotFoundException()
     return updatedUser
   }
@@ -50,6 +59,8 @@ export class UsersController {
   @Delete(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
   async remove(@Param('id') id: string) {
+    if (!isValidObjectId(id))
+      throw new BadRequestException('Incorrect id format')
     const deleteResult: any = await this.usersService.remove(id)
     if (deleteResult.deletedCount === 0) throw new NotFoundException()
   }
